expr for output neurons

+++ fppoly/fppoly.h
@@ -233,9 +233,9 @@ void update_activation_upper_bound_for_neuron(elina_manager_t *man, elina_abstra
 
 void update_activation_lower_bound_for_neuron(elina_manager_t *man, elina_abstract0_t *abs, size_t layerno, size_t neuron_no, double* coeff, size_t *dim, size_t size);
 
-elina_linexpr0_t *get_output_lexpr_defined_over_previous_layers(elina_manager_t *man, elina_abstract0_t *element, size_t neuron_no, int prev_layer);
+elina_linexpr0_t *get_output_lexpr_defined_over_previous_layers(elina_manager_t *man, elina_abstract0_t *element, int neuron_no, int prev_layer);
 
-elina_linexpr0_t *get_output_uexpr_defined_over_previous_layers(elina_manager_t *man, elina_abstract0_t *element, size_t neuron_no, int prev_layer);
+elina_linexpr0_t *get_output_uexpr_defined_over_previous_layers(elina_manager_t *man, elina_abstract0_t *element, int neuron_no, int prev_layer);
 
 #ifdef __cplusplus
  }
